requirements.txt
selenium>=4.9.0
webdriver-manager>=4.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
schedule>=1.2.0
pytest>=7.0.0
